test(snowflake): consolidate integration specs into one file
